refactor: Extract utility tools into separate module
- Created src/tools/utility_tools.py with 2 functions:
  - validate_hgvs_variant() - Mutalyzer HGVS validation & parsing
  - convert_protein_variant() - Transvar coordinate conversion
- Updated src/tools/__init__.py to export utility_tools
- Updated server.py to import and register utility tools
- Removed utility tool implementations from server.py (~60 lines)
- server.py now ~102 lines (down from ~164 lines)

Part of PR #7 in refactoring series (see REFACTORING_PLAN.md)

ALL 26 TOOLS NOW EXTRACTED INTO MODULES! 🎉
Server.py reduced from 791 to 102 lines (87% reduction)
