+++ server.py
@@ -16,7 +16,14 @@ from mcp.server.fastmcp import FastMCP
 from src.utils.api_client import fetch_marrvel_data
 
 # Import tool modules
-from src.tools import gene_tools, variant_tools, disease_tools, ortholog_tools, expression_tools
+from src.tools import (
+    gene_tools,
+    variant_tools,
+    disease_tools,
+    ortholog_tools,
+    expression_tools,
+    utility_tools
+)
 
 # Initialize FastMCP server
 mcp = FastMCP("MARRVEL")
@@ -27,6 +34,7 @@ variant_tools.register_tools(mcp)
 disease_tools.register_tools(mcp)
 ortholog_tools.register_tools(mcp)
 expression_tools.register_tools(mcp)
+utility_tools.register_tools(mcp)
 
 
 # ============================================================================
@@ -77,73 +85,11 @@ expression_tools.register_tools(mcp)
 
 
 # ============================================================================
-# UTILITY TOOLS
+# UTILITY TOOLS - Now imported from src.tools.utility_tools
 # ============================================================================
-
-@mcp.tool()
-async def validate_hgvs_variant(hgvs_variant: str) -> str:
-    """
-    Validate and parse HGVS variant nomenclature using Mutalyzer.
-    
-    Mutalyzer checks if variant descriptions are correct according to HGVS
-    nomenclature standards and provides parsed components.
-    
-    Args:
-        hgvs_variant: Variant in HGVS format
-            Examples:
-            - Genomic: "NC_000017.10:g.7577121C>T"
-            - Coding: "NM_000546.5:c.215C>G"
-            - Protein: "NP_000537.3:p.Arg72Pro"
-        
-    Returns:
-        JSON string with validation results:
-        - Validation status (valid/invalid)
-        - Parsed components
-        - Genomic coordinates
-        - Protein changes
-        - Alternative descriptions
-        
-    Example:
-        validate_hgvs_variant("NM_000546.5:c.215C>G")
-        validate_hgvs_variant("NC_000017.10:g.7577121C>T")
-    """
-    try:
-        data = await fetch_marrvel_data(f"/mutalyzer/hgvs/{hgvs_variant}")
-        return str(data)
-    except httpx.HTTPError as e:
-        return f"Error validating HGVS variant: {str(e)}"
-
-
-@mcp.tool()
-async def convert_protein_variant(protein_variant: str) -> str:
-    """
-    Convert protein-level variants to genomic coordinates using Transvar.
-    
-    Transvar is a tool for converting between different variant annotation
-    formats and coordinate systems.
-    
-    Args:
-        protein_variant: Protein variant description
-            Examples:
-            - "ENSP00000269305:p.R248Q"
-            - "NP_000537.3:p.Arg72Pro"
-        
-    Returns:
-        JSON string with converted coordinates:
-        - Genomic coordinates (hg19, hg38)
-        - cDNA changes
-        - Multiple transcript mappings
-        - Alternative annotations
-        
-    Example:
-        convert_protein_variant("ENSP00000269305:p.R248Q")
-        convert_protein_variant("NP_000537.3:p.Arg72Pro")
-    """
-    try:
-        data = await fetch_marrvel_data(f"/transvar/protein/{protein_variant}")
-        return str(data)
-    except httpx.HTTPError as e:
-        return f"Error converting protein variant: {str(e)}"
+# The following utility tools are now registered from the utility_tools module:
+# - validate_hgvs_variant (Mutalyzer HGVS validation)
+# - convert_protein_variant (Transvar coordinate conversion)
 
 
 # ============================================================================
